requirements.txt
Jinja2==3.1.2
jsonschema==4.16.0
numpy==1.23.3
pandas==1.4.4
Pillow==9.2.0
requests==2.28.1
urllib3==1.26.12
